разделил main и создал папку src с файлами отдельно под каждый класс и отдельный файл для utils
