+++ src/product.py
@@ -1,0 +1,12 @@
+class Product:
+    """Класс для предоставления продуктов"""
+    title: str
+    description: str
+    price: float
+    quantity_in_stock: int
+
+    def __init__(self, title, description, price, quantity_in_stock):
+        self.title = title
+        self.description = description
+        self.price = price
+        self.quantity_in_stock = quantity_in_stock
